refactor(dispatch): guard against nil params
One of our tests has brought up a funny interaction! In `TestProfileRequestSave`, we were pass in `nil` parameters. Then, `Dispatch` casts those `nil` parameters (a `nil` pointer to a `*config.ProfilePod`), to an `interface{}`.

So it's as if we have done this:
```
var param *config.ProfilePod = nil
var i interface{}
i = param
```

Here is the expectation:
```
param == nil // true
i == nil // true
```

BUT that's not the case, because of how interfaces are defined. An interface consists of a type and a value, both must be nil for the interface to be considered nil. But in the above we have set interface i to be `(*config.ProfilePod)(nil)`, so we can't just rely on `i == nil` to guard against nil calling.

Where is this causing us a problem? In `Dispatch`, we assert that the params are of type `Validator` in order to validate them before we pass them to the method. *config.ProfilePod IS a validator.  But! we then call `param.Validate()`, which tries to automatically dereference the pointer, and the pointer is nil! Which causes a nil pointer dereference panic.

So we should guard against that! Using reflect we can ensure if we have a param that is is nil, or a param that is both a pointer & is nil, we return w/ an error.

At the end of the day... it turns out we specifically DON'T want to validate the `config.ProfilePod` that comes in via the Save path, because the user can just be updating a single part of the config & that config won't contains a private key, that the validator expects. So for this particular case, it's actually a moot point. But, it seems really easy for future us to pass in a `nil` params to `Dispatch` & cause a panic. This change guards future foot guns!